Test for string length.
Strings currently can be only 255 characters long. Any
characters that go over, are cut off.

Tests will fail after this commit. Translations will need
to be edited or Textpattern's field type for the
translation changed. There are ton of strings that
go over this limit. Translating can add length in some
cases. Not all languages are as short as in English.