Unifying welcome email - remove full stop/period from you_have_been_registered string
Lessens the awkwardness of this incorrectly-formatted sentence in the
welcome email:

```
You have been registered as a contributor to the site. My Site Name
```

Some textpacks had a trailing . after the `you_have_been_registered`
string, others didn’t - this unifies them, to the best of my knowledge.

It still doesn’t have a trailing . to indicate the end of the sentence
proper (after site name), but it’s a start.

The string `you_have_been_registered string` is currently only used in
the welcome email.